Renaming OICDAuthOptions to DefaultAuthOptions, since there is nothing on oidc in the clazz
