+++ auth/src/main/java/org/aerogear/mobile/auth/AuthService.java
@@ -7,10 +7,10 @@ import android.support.annotation.NonNull;
 import org.aerogear.mobile.auth.authenticator.AuthorizationServiceFactory;
 import org.aerogear.mobile.auth.configuration.AuthServiceConfiguration;
 import org.aerogear.mobile.auth.configuration.KeycloakConfiguration;
-import org.aerogear.mobile.auth.authenticator.OIDCAuthenticateOptions;
+import org.aerogear.mobile.auth.authenticator.DefaultAuthenticateOptions;
 import org.aerogear.mobile.auth.credentials.JwksManager;
 import org.aerogear.mobile.auth.credentials.OIDCCredentials;
-import org.aerogear.mobile.auth.authenticator.OIDCAuthenticatorImpl;
+import org.aerogear.mobile.auth.authenticator.oidc.OIDCAuthenticatorImpl;
 import org.aerogear.mobile.auth.user.UserPrincipal;
 import org.aerogear.mobile.auth.utils.UserIdentityParser;
 import org.aerogear.mobile.core.MobileCore;
@@ -135,7 +135,7 @@ public class AuthService implements ServiceModule {
      * @param authOptions the authentication options
      * @param callback the callback function that will be invoked with the user info
      */
-    public void login(@NonNull final OIDCAuthenticateOptions authOptions, @NonNull final Callback<UserPrincipal> callback) {
+    public void login(@NonNull final DefaultAuthenticateOptions authOptions, @NonNull final Callback<UserPrincipal> callback) {
         failIfNotReady();
         oidcAuthenticatorImpl.authenticate(authOptions, callback);
     }
